Added a Kafka+SQL storage variant
