+++ app/src/main/resources/application.properties
@@ -5,12 +5,13 @@ quarkus.index-dependency.jaxrs.artifact-id=jboss-jaxrs-api_2.1_spec
 
 # === Dev profile - see README
 %dev.quarkus.http.port=${HTTP_PORT:8080}
-%dev.quarkus.log.level=${LOG_LEVEL:DEBUG}
+%dev.quarkus.log.level=${LOG_LEVEL:INFO}
+%dev.quarkus.log.category."io.apicurio".level=${REGISTRY_LOG_LEVEL:DEBUG}
 %dev.quarkus.log.console.enable=true
 %dev.quarkus.datasource.db-kind=postgresql
-%dev.quarkus.datasource.jdbc.url=${QUARKUS_DATASOURCE_URL:jdbc:postgresql://localhost:5432/registry}
-%dev.quarkus.datasource.username=${QUARKUS_DATASOURCE_USERNAME:postgres}
-%dev.quarkus.datasource.password=${QUARKUS_DATASOURCE_PASSWORD:postgres}
+%dev.quarkus.datasource.jdbc.url=${REGISTRY_DATASOURCE_URL:jdbc:postgresql://localhost:5432/registry}
+%dev.quarkus.datasource.username=${REGISTRY_DATASOURCE_USERNAME:postgres}
+%dev.quarkus.datasource.password=${REGISTRY_DATASOURCE_PASSWORD:postgres}
 %dev.registry.sql.init=true
 %dev.quarkus.hibernate-orm.log.sql=true
 %dev.quarkus.hibernate-orm.log.jdbc-warnings=true
@@ -58,6 +59,7 @@ quarkus.index-dependency.jaxrs.artifact-id=jboss-jaxrs-api_2.1_spec
 # Log
 %prod.quarkus.log.level=${LOG_LEVEL:INFO}
 %prod.quarkus.log.console.enable=true
+%prod.quarkus.log.category."io.apicurio".level=${REGISTRY_LOG_LEVEL:INFO}
 
 # SQL
 %prod.quarkus.datasource.db-kind=postgresql
