+++ app/src/test/resources/application.properties
@@ -1,12 +1,14 @@
 # Test config
 
 %test.quarkus.log.level=${LOG_LEVEL:INFO}
+%test.quarkus.log.category."io.apicurio".level=${REGISTRY_LOG_LEVEL:DEBUG}
 %test.quarkus.log.console.enable=true
 
 %test.quarkus.datasource.db-kind=h2
 %test.quarkus.datasource.min-size=3
 %test.quarkus.datasource.max-size=12
-%test.quarkus.datasource.jdbc.url=jdbc:h2:tcp://localhost:9123/mem:registry;DB_CLOSE_DELAY=-1;IFEXISTS=FALSE;
+#%test.quarkus.datasource.jdbc.url=jdbc:h2:tcp://localhost:9123/mem:registry;DB_CLOSE_DELAY=-1;IFEXISTS=FALSE;
+%test.quarkus.datasource.jdbc.url=jdbc:h2:mem:registry_db
 %test.registry.sql.init=true
 h2.jar.file.path=${project.parent.basedir}/storage/sql/target/h2.jar
 h2.port=9123
@@ -23,6 +25,11 @@ h2.port=9123
 %test.registry.kafka.snapshot-producer.retries=3
 %test.registry.kafka.snapshot-producer.acks=all
 %test.registry.kafka.snapshot.requests=10
+
+# KSQL
+%test.registry.ksql.bootstrap.servers=localhost:9092
+%test.registry.ksql.topic=ksql-journal
+%test.registry.ksql.consumer.startupLag=500
 
 # Streams -- when used
 %test.registry.streams.common.bootstrap.servers=${bootstrap.servers:localhost:9092}
