+++ distro/docker/src/main/docker/Dockerfile.ksql.jvm
@@ -1,0 +1,9 @@
+FROM fabric8/java-alpine-openjdk8-jre
+
+ENV APP_URL="${docker.ksql.file}"
+ENV JAVA_OPTIONS="-Dquarkus.http.host=0.0.0.0 -Djava.util.logging.manager=org.jboss.logmanager.LogManager"
+ENV AB_ENABLED=jmx_exporter
+
+ADD "$APP_URL" /deployments
+
+ENTRYPOINT [ "/deployments/run-java.sh" ]
