+++ storage/ksql/src/main/java/io/apicurio/registry/storage/impl/ksql/JournalRecordArgument.java
@@ -1,0 +1,63 @@
+/*
+ * Copyright 2020 Red Hat
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.apicurio.registry.storage.impl.ksql;
+
+import com.fasterxml.jackson.databind.JsonNode;
+
+/**
+ * @author eric.wittmann@gmail.com
+ */
+public class JournalRecordArgument {
+    
+    private String className;
+    private JsonNode value;
+    
+    /**
+     * Constructor.
+     */
+    public JournalRecordArgument() {
+    }
+
+    /**
+     * @return the value
+     */
+    public JsonNode getValue() {
+        return value;
+    }
+
+    /**
+     * @param value the value to set
+     */
+    public void setValue(JsonNode value) {
+        this.value = value;
+    }
+
+    /**
+     * @return the className
+     */
+    public String getClassName() {
+        return className;
+    }
+
+    /**
+     * @param className the className to set
+     */
+    public void setClassName(String className) {
+        this.className = className;
+    }
+
+}
