+++ storage/ksql/src/main/java/io/apicurio/registry/storage/impl/ksql/JournalRecordDeserializer.java
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2020 Red Hat
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.apicurio.registry.storage.impl.ksql;
+
+import java.io.IOException;
+
+import org.apache.kafka.common.serialization.Deserializer;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+/**
+ * @author eric.wittmann@gmail.com
+ */
+public class JournalRecordDeserializer implements Deserializer<JournalRecord> {
+    
+    private ObjectMapper mapper = new ObjectMapper();
+
+    /**
+     * @see org.apache.kafka.common.serialization.Deserializer#deserialize(java.lang.String, byte[])
+     */
+    @Override
+    public JournalRecord deserialize(String topic, byte[] data) {
+        try {
+            return mapper.readValue(data, JournalRecord.class);
+        } catch (IOException e) {
+            // TODO figure something out here
+            e.printStackTrace();
+            return null;
+        }
+    }
+
+}
