+++ storage/ksql/src/main/resources/overlay.properties
@@ -1,0 +1,19 @@
+%dev.quarkus.datasource.db-kind=h2
+%dev.quarkus.datasource.jdbc.url=${REGISTRY_DATASOURCE_URL:jdbc:h2:mem:registry_db}
+%dev.quarkus.datasource.username=${REGISTRY_DATASOURCE_USERNAME:sa}
+%dev.quarkus.datasource.password=${REGISTRY_DATASOURCE_PASSWORD:sa}
+
+%dev.registry.ksql.bootstrap.servers=localhost:9092
+%dev.registry.ksql.topic=ksql-journal
+%dev.registry.ksql.consumer.startupLag=100
+%dev.registry.ksql.consumer.poll.timeout=1000
+
+%prod.quarkus.datasource.db-kind=h2
+%prod.quarkus.datasource.jdbc.url=${REGISTRY_DATASOURCE_URL:jdbc:h2:mem:registry_db}
+%prod.quarkus.datasource.username=${REGISTRY_DATASOURCE_USERNAME:sa}
+%prod.quarkus.datasource.password=${REGISTRY_DATASOURCE_PASSWORD:sa}
+
+%prod.registry.ksql.bootstrap.servers=localhost:9092
+%prod.registry.ksql.topic=ksql-journal
+%prod.registry.ksql.consumer.startupLag=100
+%prod.registry.ksql.consumer.poll.timeout=100
