+++ storage/sql/src/test/resources/application.properties
@@ -1,10 +1,12 @@
 # Test config
 
 %test.quarkus.log.level=${LOG_LEVEL:INFO}
+%test.quarkus.log.category."io.apicurio".level=${REGISTRY_LOG_LEVEL:DEBUG}
 %test.quarkus.log.console.enable=true
 
 %test.quarkus.datasource.db-kind=h2
-%test.quarkus.datasource.jdbc.url=jdbc:h2:tcp://localhost:9123/mem:registry;DB_CLOSE_DELAY=-1;IFEXISTS=FALSE;
+%test.quarkus.datasource.jdbc.url=jdbc:h2:mem:registry_db
+#%test.quarkus.datasource.jdbc.url=jdbc:h2:tcp://localhost:9123/mem:registry;DB_CLOSE_DELAY=-1;IFEXISTS=FALSE;
 
 #%test.quarkus.datasource.db-kind=postgresql
 #%test.quarkus.datasource.jdbc.url=jdbc:postgresql://127.0.0.1:5432/registry
