+++ utils/sql/src/main/java/io/apicurio/registry/storage/impl/sql/CommonSqlStatements.java
