+++ utils/sql/src/main/java/io/apicurio/registry/storage/impl/sql/H2SqlStatements.java
