+++ utils/sql/src/main/java/io/apicurio/registry/storage/impl/sql/mappers/SearchedArtifactMapper.java
