+++ utils/sql/src/main/resources/io/apicurio/registry/storage/impl/sql/h2.ddl
