+++ utils/sql/src/main/resources/io/apicurio/registry/storage/impl/sql/postgresql.ddl
