+++ utils/sql/src/test/java/io/apicurio/registry/storage/impl/sql/SqlUtilTest.java
